fully replaced the original macos implementation
