+++ src/backend/iohidmanager/mod.rs
@@ -1,174 +1,233 @@
-mod device;
-mod manager;
-mod runloop;
-mod service;
-mod utils;
+mod device_info;
+mod read_writer;
 
-use std::sync::Arc;
-
-use core_foundation::array::CFArray;
-use core_foundation::base::TCFType;
-use core_foundation::dictionary::CFDictionary;
-use core_foundation::runloop::{kCFRunLoopDefaultMode, CFRunLoop};
-use futures_lite::stream::iter;
-use futures_lite::StreamExt;
-use core_foundation::string::CFString;
-use io_kit_sys::types::IOOptionBits;
-use crate::backend::iohidmanager::device::{AsyncReportReader, IOHIDDevice};
-
-use io_kit_sys::hid::keys::*;
-use crate::backend::iohidmanager::utils::{CFDictionaryExt};
-use crate::backend::iohidmanager::runloop::RunLoop;
-use crate::backend::iohidmanager::service::{IOService, RegistryEntryId};
-use crate::{AsyncHidRead, AsyncHidWrite, DeviceId, DeviceInfo, HidError, HidResult};
-use crate::backend::iohidmanager::manager::IOHIDManager;
+use crate::backend::iohidmanager::device_info::{get_device_id, get_device_info};
+use crate::backend::iohidmanager::read_writer::DeviceReadWriter;
 use crate::backend::{Backend, DeviceInfoStream};
 use crate::utils::TryIterExt;
+use crate::{ensure, DeviceEvent, DeviceId, HidError, HidResult};
+use dispatch2::{DispatchQueue, DispatchQueueAttr, DispatchRetained};
+use futures_lite::stream::{iter, Boxed};
+use futures_lite::{Stream, StreamExt};
+use objc2_core_foundation::{CFDictionary, CFRetained};
+use objc2_io_kit::{kIOMasterPortDefault, kIOReturnSuccess, IOHIDDevice, IOHIDManager, IOHIDManagerOptions, IORegistryEntryIDMatching, IOReturn, IOServiceGetMatchingService};
+use std::ffi::c_void;
+use std::pin::Pin;
+use std::ptr::{null, NonNull};
+use std::sync::{Arc, LazyLock, Mutex, Once};
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::task::{Context, Poll};
+use atomic_waker::AtomicWaker;
+use block2::RcBlock;
+use crossbeam_queue::ArrayQueue;
+use log::{debug, trace};
 
-#[derive(Default)]
-pub struct IoHidManagerBackend;
+static DISPATCH_QUEUE: LazyLock<DispatchRetained<DispatchQueue>> = LazyLock::new(|| DispatchQueue::new("async-hid", DispatchQueueAttr::SERIAL));
+
+pub type IoHidManagerBackend = Arc<IoHidManagerBackendInner>;
+
+pub struct IoHidManagerBackendInner{
+    manager: CFRetained<IOHIDManager>,
+    callback_context: *const ManagerCallbackContext,
+}
+
+//SAFETY: IOHIDManager is immediately connected to a dispatch queue, and 
+// all functions are called on that queue to the best of my knowledge
+unsafe impl Send for IoHidManagerBackendInner {}
+unsafe impl Sync for IoHidManagerBackendInner {}
+
+impl Default for IoHidManagerBackendInner {
+    fn default() -> Self {
+        unsafe {
+            trace!("Creating manager");
+            let manager = IOHIDManager::new(None, IOHIDManagerOptions::None.bits());
+            manager.set_dispatch_queue(&*DISPATCH_QUEUE);
+
+            let context = Box::into_raw(Box::new(ManagerCallbackContext::default()));
+            manager.set_device_matching(None);
+            manager.register_device_matching_callback(Some(ManagerCallbackContext::added_callback), context as *mut c_void);
+            manager.register_device_removal_callback(Some(ManagerCallbackContext::removed_callback), context as *mut c_void);
+            trace!("Scheduling manager with run loop");
+            manager.activate();
+
+            Self {
+                manager,
+                callback_context: context as _,
+            }
+        }
+    }
+}
+
+impl IoHidManagerBackendInner {
+    fn callback_context(&self) -> &ManagerCallbackContext {
+        unsafe { &*self.callback_context }
+    }
+}
+
+impl Drop for IoHidManagerBackendInner {
+    fn drop(&mut self) {
+        unsafe {
+            let once = Arc::new(Once::new());
+            let block = RcBlock::new({
+                let once = once.clone();
+                move || once.call_once(|| trace!("Finished canceling manager"))
+            });
+
+            self.manager.set_cancel_handler(RcBlock::as_ptr(&block));
+            self.manager.cancel();
+            trace!("Waiting for manager cancel to finish");
+            once.wait();
+            trace!("Resuming destructor of manager");
+
+            //SAFETY: Cancel will ensure that the callbacks are never called again, allowing us to free the data structures reference in them
+            drop(Box::from_raw(self.callback_context as *mut ManagerCallbackContext));
+            self.callback_context = null();
+        }
+    }
+}
 
 impl Backend for IoHidManagerBackend {
-    type Reader = InputReceiver;
-    type Writer = Arc<BackendDevice>;
+    type Reader = Arc<DeviceReadWriter>;
+    type Writer = Arc<DeviceReadWriter>;
 
     async fn enumerate(&self) -> HidResult<DeviceInfoStream> {
-        let mut manager = IOHIDManager::new()?;
-        let devices = manager
-            .get_devices()?
-            .into_iter()
-            .map(get_device_infos)
-            .try_flatten();
+        let device_infos = unsafe {
+            let device_set = self.manager.devices().expect("Failed to get devices");
+            let mut devices: Vec<NonNull<IOHIDDevice>> = vec![NonNull::dangling(); device_set.count() as usize];
+            device_set.values(devices.as_mut_ptr().cast());
+            devices
+                .iter()
+                .map(|d| get_device_info(d.as_ref()))
+                .try_flatten()
+                .collect::<Vec<_>>()
+        };
+        Ok(iter(device_infos).boxed())
+    }
 
-        Ok(iter(devices).boxed())
+    fn watch(&self) -> HidResult<Boxed<DeviceEvent>> {
+        let watcher = DeviceWatcher::new(self.clone());
+        Ok(watcher.boxed())
     }
 
     async fn open(&self, id: &DeviceId, read: bool, write: bool) -> HidResult<(Option<Self::Reader>, Option<Self::Writer>)> {
         let id = match id {
-            DeviceId::RegistryEntryId(id) => RegistryEntryId(*id)
+            DeviceId::RegistryEntryId(id) => *id
         };
-        let open_options = 0;
-        let device = IOHIDDevice::try_from(id)?;
-        device.open(open_options)?;
-        let device = Arc::new(BackendDevice {
-            device,
-            open_options,
-        });
-        
-        let input_receiver = if read {
-            Some(InputReceiver::new(device.clone()).await?)
-        } else {
-            None
+        let device = unsafe {
+            let service = IOServiceGetMatchingService(kIOMasterPortDefault, IORegistryEntryIDMatching(id).map(|d|d.downcast::<CFDictionary>().unwrap()));
+            ensure!(service != 0, HidError::NotConnected);
+            let device = IOHIDDevice::new(None, service).ok_or(HidError::message("Failed to create device"))?;
+            device.set_dispatch_queue(&*DISPATCH_QUEUE);
+            ensure!(device.open(DeviceReadWriter::DEVICE_OPTIONS) == kIOReturnSuccess, HidError::message("Failed to open device"));
+            device
         };
+        let rw = Arc::new(DeviceReadWriter::new(device, read, write)?);
+        Ok((read.then_some(rw.clone()), write.then_some(rw)))
+    }
 
-        Ok((input_receiver, write.then_some(device)))
+
+}
+
+pub struct DeviceWatcher {
+    id: u64,
+    queue: Arc<AsyncQueue<DeviceEvent>>,
+    backend: IoHidManagerBackend,
+}
+
+impl DeviceWatcher {
+    pub fn new(backend: IoHidManagerBackend) -> Self {
+        let (id, queue) = backend.callback_context().register_watcher();
+        Self {
+            id,
+            queue,
+            backend
+        }
+    }
+}
+
+impl Stream for DeviceWatcher {
+    type Item = DeviceEvent;
+
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        self.queue.poll_next(cx)
+    }
+}
+
+impl Drop for DeviceWatcher {
+    fn drop(&mut self) {
+        self.backend.callback_context().unregister_watcher(self.id);
     }
 }
 
 
-fn get_device_infos(device: IOHIDDevice) -> HidResult<Vec<DeviceInfo>> {
-    let primary_usage_page = device.get_i32_property(kIOHIDPrimaryUsagePageKey)? as u16;
-    let primary_usage = device.get_i32_property(kIOHIDPrimaryUsageKey)? as u16;
-    let vendor_id = device.get_i32_property(kIOHIDVendorIDKey)? as u16;
-    let product_id = device.get_i32_property(kIOHIDProductIDKey)? as u16;
-    let serial_number = device.get_string_property(kIOHIDProductKey).ok();
-    let name = device.get_string_property(kIOHIDProductKey)?;
-    let id = IOService::try_from(&device).and_then(|i| i.get_registry_entry_id())?;
-
-    let info = DeviceInfo {
-        id: DeviceId::RegistryEntryId(id.0),
-        name,
-        product_id,
-        vendor_id,
-        usage_id: primary_usage,
-        usage_page: primary_usage_page,
-        serial_number,
-    };
-
-    let mut results = Vec::new();
-    results.extend(
-        device
-            .property::<CFArray>(kIOHIDDeviceUsagePairsKey)?
-            .iter()
-            .map(|i| unsafe { CFDictionary::wrap_under_get_rule(*i as _) })
-            .filter_map(|dict| {
-                let usage = dict.lookup_i32(kIOHIDDeviceUsageKey).ok()? as u16;
-                let usage_page = dict.lookup_i32(kIOHIDDeviceUsagePageKey).ok()? as u16;
-                Some((usage, usage_page))
-            })
-            .filter(|(usage, usage_page)| (*usage_page != primary_usage_page) || (*usage != primary_usage))
-            .map(|(usage_id, usage_page)| DeviceInfo {
-                usage_id,
-                usage_page,
-                ..info.clone()
-            })
-    );
-    results.push(info.clone());
-
-    Ok(results)
+#[derive(Default)]
+struct ManagerCallbackContext {
+    next_id: AtomicU64,
+    watchers: Mutex<Vec<(u64, Arc<AsyncQueue<DeviceEvent>>)>>
 }
 
-pub struct InputReceiver {
-    device: Arc<BackendDevice>,
-    run_loop: Arc<RunLoop>,
-    reader: AsyncReportReader,
-}
-
-impl InputReceiver {
-    async fn new(device: Arc<BackendDevice>) -> HidResult<Self> {
-        let reader = AsyncReportReader::new(&device.device)?;
-        let run_loop = RunLoop::get_run_loop().await?;
-        run_loop.schedule_device(&device.device)?;
-
-        Ok(Self {
-            device,
-            run_loop,
-            reader
-        })
-    }
-
-    fn stop(&self, device: &IOHIDDevice) {
-        self.run_loop
-            .unschedule_device(device)
-            .unwrap_or_else(|_| log::warn!("Failed to unschedule IOHIDDevice from run loop"));
-        let default_mode = unsafe { CFString::wrap_under_create_rule(kCFRunLoopDefaultMode) };
-        device.schedule_with_runloop(&CFRunLoop::get_main(), &default_mode);
+impl ManagerCallbackContext {
+    
+    pub fn register_watcher(&self) -> (u64, Arc<AsyncQueue<DeviceEvent>>) {
+        let id = self.next_id.fetch_add(1, Ordering::Relaxed);
+        let queue = Arc::new(AsyncQueue::new(64));
+        let mut watchers = self.watchers.lock().unwrap();
+        watchers.push((id, queue.clone()));
+        trace!("Registered watcher with id {} (total: {})", id, watchers.len());
+        (id, queue)
     }
     
-}
+    pub fn unregister_watcher(&self, id: u64) {
+        let mut watchers = self.watchers.lock().unwrap();
+        watchers.retain(|&(i, _)| i != id);
+        trace!("Unregistered watcher with id {} (remaining: {})", id, watchers.len());
+    }
+    
+    fn notify_watchers(&self, event: DeviceEvent) {
+        let mut watchers = self.watchers.lock().unwrap();
+        for (_, queue) in watchers.iter_mut() {
+            queue.force_push(event.clone());
+        }
+    }
+    
+    unsafe extern "C-unwind" fn added_callback(context: *mut c_void, _result: IOReturn, _sender: *mut c_void, device: NonNull<IOHIDDevice>) {
+        let this: &Self = &*(context as *const Self);
+        match get_device_id(device.as_ref()) {
+            Ok(id) => this.notify_watchers(DeviceEvent::Connected(id)),
+            Err(err) => debug!("Failed to get device id: {}", err)
+        }
+    }
 
-impl AsyncHidRead for InputReceiver {
-    async fn read_input_report<'a>(&'a mut self, buf: &'a mut [u8]) -> HidResult<usize> {
-        self.reader.read(buf).await
+    unsafe extern "C-unwind" fn removed_callback(context: *mut c_void, _result: IOReturn, _sender: *mut c_void, device: NonNull<IOHIDDevice>) {
+        let this: &Self = &*(context as *const Self);
+        match get_device_id(device.as_ref()) {
+            Ok(id) => this.notify_watchers(DeviceEvent::Disconnected(id)),
+            Err(err) => debug!("Failed to get device id: {}", err)
+        }
     }
 }
 
-impl Drop for InputReceiver {
-    fn drop(&mut self) {
-        self.stop(&self.device.device)
-    }
+pub struct AsyncQueue<T> {
+    items: ArrayQueue<T>,
+    waker: AtomicWaker
 }
 
-pub struct BackendDevice {
-    device: IOHIDDevice,
-    open_options: IOOptionBits
-}
-
-impl Drop for BackendDevice {
-    fn drop(&mut self) {
-        self.device
-            .close(self.open_options)
-            .err()
-            .filter(|err| !matches!(*err, HidError::Disconnected))
-            .inspect(|err| log::warn!("Failed to close IOHIDDevice\n\t{err:?}"));
+impl<T> AsyncQueue<T> {
+    pub fn new(cap: usize) -> Self {
+        Self {
+            items: ArrayQueue::new(cap),
+            waker: AtomicWaker::new()
+        }
     }
-}
-
-impl AsyncHidWrite for Arc<BackendDevice> {
-    async fn write_output_report<'a>(&'a mut self, buf: &'a [u8]) -> HidResult<()> {
-        let report_id = buf[0];
-        let data_to_send = if report_id == 0x0 { &buf[1..] } else { buf };
-
-        self.device.set_report(1, report_id as _, data_to_send)
+    
+    pub fn force_push(&self, item: T) {
+        self.items.force_push(item);
+        self.waker.wake();
     }
+    
+    pub fn poll_next(&self, cx: &mut Context<'_>) -> Poll<Option<T>> {
+        self.waker.register(cx.waker());
+        self.items.pop().map(Some).map(Poll::Ready).unwrap_or(Poll::Pending)
+    }
+    
 }
